Update batchHathiTrust.rb
Enable multiple HathiTrust items to be combined into a single batch for submission.

+++ lib/batchHathiTrust.rb
@@ -15,16 +15,11 @@
 # will zip the package into a ZIP file with a filename meeting
 # HathiTrust requirements.
 #
-# The job expects TIFF filenames which contain a unique number
-# ranging from 1 to however many images are contained in the scan.
-# The number can contain leading zeroes and must be followed by
-# filename extension, or by '_L' or '_R' and then the filename
-# extension. The filename extension can be .tif or .tiff. The job
-# ignores everything in the filename preceeding the page numeric
-# value. Note that the numbers represent scans, not page numbers,
-# and should include front and back covers, endpapers, blank pages,
-# and front and back matter. Scans can be left-to-right or
-# right-to-left.
+# The job expects TIFF files with a filename extension of .tif,
+# .tiff, .TIF, or .TIFF. The job sorts the files alphanumerically
+# then renames them 00000001.tif, etc. The scans should include
+# front and back covers, endpapers, blank  pages, and front and
+# back matter. Scans can be left-to-right or right-to-left.
 class BatchHathiTrust < Batch
   require 'fileutils'
   require 'digest'
@@ -43,14 +38,15 @@ class BatchHathiTrust < Batch
     $ui.message("\nYou are about to package material for submission to HathiTrust." +
                 "\n  See https://www.hathitrust.org/member-libraries/contribute-content/" +
                 "\n  for details on submission requirements." +
-                "\n\nIn particular, a Digital Asset Submission Inventory form and a" +
-                "\n  an Administrative Coversheet form must be delivered to HathiTrust" +
+                "\n\nIn particular, a Digital Asset Submission Inventory form and an" +
+                "\n  Administrative Coversheet form must be delivered to HathiTrust" +
                 "\n  before any submission." +
-                "\n\nPlease provide a directory containing the TIFF files and a MARC record" +
-                "\n  for submission. The MARC record must be for the PRINT version of the item" +
-                "\n  and should contain the local record identifier in 001 and the OCLC number" +
-                "\n  in 035. Each volume of a multi-part title must be submitted to HathiTrust" +
-                "\n  separately. Please have the barcode for the item ready.")
+                "\n\nPlease provide a directory containing the TIFF files and a MARC" +
+                "\n  record for submission. The MARC record must be for the PRINT version" +
+                "\n  of the item and should contain the local record identifier in 001 and" +
+                "\n  the OCLC number in 035. Each volume of a multi-part title must be" +
+                "\n  submitted to HathiTrust separately. Please have the barcode or Internet" +
+                "\n  Archive ARK number for the item ready.")
 
     @user_directory = $ui.question("\nWhat is the directory you are working with?")
     while !Dir.exist?(@user_directory)
@@ -59,12 +55,16 @@ class BatchHathiTrust < Batch
     end
 
     Dir.chdir(@user_directory)
-    files = Dir.glob('**/*.mrc')
+    files = Dir.glob('**/*.{mrc,marc}')
     if files.none?
-      $ui.splash('No MARC files in directory')
+      $ui.splash('No MARC files found in directory; exiting')
       return
     end
-    reader = MARC::Reader.new(files[0])
+    if files.length > 0
+      $ui.splash("More than one MARC file found--using first file:\n  " + files[0])
+    end
+    f = File.open(files[0])
+    reader = MARC::Reader.new(f)
     oclc = ''
     record = reader.first
     rectype = record.leader[6,1]
@@ -79,12 +79,30 @@ class BatchHathiTrust < Batch
       recform = record['008'].value[23]
     end
     if recform == 'o'
-      $ui.splash('Use a MARC record for the print version rather than the online version; exiting')
+      $ui.splash("Record is for an online version\nPlease use a MARC record for the print version; exiting")
       return
     elsif recform != 'd' and recform != 'r' and recform != '|' and recform != ' '
       $ui.splash('MARC record is not for a print resource; exiting')
       return
     end
+    title_field = record['245']
+    msg = title_field['a']
+	if title_field['b']
+      msg << " " + title_field['b']
+    end
+	if title_field['n']
+      msg << " " + title_field['n']
+    end
+	if title_field['p']
+      msg << " " + title_field['p']
+    end
+	if title_field['c']
+      msg << " " + title_field['c']
+    end
+    $ui.message("Title is: " + msg)
+
+    # find first OCLC number
+    oclc = ''
     record.each_by_tag('035') { |field|
       value = field['a'].to_s
       if value[0,7] == "(OCoLC)" and oclc == ''
@@ -95,54 +113,100 @@ class BatchHathiTrust < Batch
       $ui.splash('MARC record does not contain an OCLC number in 035; exiting')
       return
     end
-
+    $ui.message("OCLC number is: " + oclc.to_s)
+    if !record['003']
+      field = MARC::ControlField.new('003', 'MMeT')
+      record << field
+    end
+    # Check OCLC number in HathiTrust
     uri = URI('https://catalog.hathitrust.org/api/volumes/full/oclc/' + oclc + '.json')
     response = Net::HTTP.get(uri)
-    if JSON[response]["records"].count != 0
+    if JSON[response]["records"].length != 0
       $ui.splash('This OCLC number is already in HathiTrust; exiting')
       return
     else
-      $ui.splash('This is NOT in HathiTrust -- creating package directory')
+      $ui.splash('This is NOT in HathiTrust--continuing')
     end
-
-    @copy_of_directory = @user_directory + '_Packaged'
-    if Dir.exist?(@copy_of_directory) then FileUtils.remove_dir(@copy_of_directory) end
-    FileUtils.mkdir_p @copy_of_directory
-    $ui.debug()
 
     id = $ui.question("\nWhat is the barcode or ARK ID of this submission?")
     if id.start_with? "ark" then
       id = id.gsub("+", ":").gsub("=", "/")
     end
-
-    if !record['003']
-      field = MARC::ControlField.new('003', 'MMeT')
-      record << field
-    end
-    field = MARC::DataField.new('955', '0', '0',
-      MARC::Subfield.new('b', id))
-    if $ui.yesno('Is this a multi-volume MARC record?', 'n')
+    field = MARC::DataField.new('955', '0', '0', MARC::Subfield.new('b', id))
+    volume = 0
+    if $ui.yesno("\nIs this a multi-volume MARC record?", 'n')
       volume = $ui.question('What volume is this submission?')
       field.append(MARC::Subfield.new('v', volume))
     end
     record << field
-    xmlname = File.join(@copy_of_directory, id.gsub(":", "+").gsub("/", "=") + '.xml')
-    MARC::XMLWriter.new(xmlname) { |w| w.write(record) }
+    f.close()
 
-    $ui.message('Copying images')
-    Dir.glob('**/*.tif').each { |f|
-	  imagenum = f[/(\d+)(_[LR])?\.tiff?/,1]
-      newfile = File.join(@copy_of_directory, sprintf("%08d.tif", imagenum.to_i))
-      FileUtils.cp(f, newfile)
-    }
-    Dir.chdir(@copy_of_directory)
+    batch = $ui.yesno("\nWill this submission be part of a batch submission (multiple objects)?", 'y')
+    if batch
+      $ui.message("\nEnter the name of the batch submission directory.")
+      $ui.message("\nIf the directory does not already exist, it will be created.")
+      d = $ui.question('', 'Batch directory: ')
+      while d == ""
+        $ui.message("\nEntry was blank; please enter the name of the batch submission directory.")
+        d = $ui.question('', 'Batch directory: ').strip
+      end
+      batch_directory = File.expand_path(d, File.dirname(@user_directory))
+      batch_xml = File.join(batch_directory, File.basename(batch_directory) + ".xml")
+      if File.exist?(batch_xml)
+        $ui.message("Checking for duplicate submission.")
+	    f = File.open(batch_xml)
+        reader = MARC::XMLReader.new(f, parser: "nokogiri")
+        reader.each do |batchrecord|
+          htfield = batchrecord['955']
+          htid = htfield['b']
+          if htid == id
+            $ui.splash('This volume (id: ' + id + ') already exists in the batch package; exiting')
+            return
+          end
+        end
+        f.close()
+      end
+    end
 
-    # Create OCR files for each page
+    # generating YAML data
+    imagefiles = Dir.glob('**/*.{tif,tiff,TIF,TIFF}')
+    if imagefiles.none?
+      $ui.splash('No TIFF image files found; exiting')
+      return
+    end
+    $ui.message(imagefiles.length.to_s + " images in package\n")
+    exif = EXIFR::TIFF.new(imagefiles[0])
+    ymltext = "scanner_user: Tufts University, Tisch Library Digital Initiatives Dept.\n"
+    if exif.date_time and exif.date_time != ""
+	  $ui.message "EXIF date_time: " + exif.date_time.to_s
+      ymltext << 'capture_date: ' + exif.date_time.strftime("%Y-%m-%dT%H:%M:%S%:z") + "\n"
+    elsif
+      digidate = $ui.question("What day was this item digitized (mm/dd/yyyy)?")
+      datetime = Time.strptime(digidate, "%m/%d/%Y")
+	  ymltext << 'capture_date: ' + datetime.strftime("%Y-%m-%dT%H:%M:%S%:z") + "\n"
+    end
+    if exif.make and exif.make != ""
+      ymltext << 'scanner_make: ' + exif.make + "\n"
+    end
+    if exif.model and exif.model != ""
+      ymltext << 'scanner_model: ' + exif.model + "\n"
+    end
+    if $ui.yesno("\nWas the resource scanned left-to-right?", 'y')
+      ymltext << "scanning_order: left-to-right\n"
+    else
+      ymltext << "scanning_order: right-to-left\n"
+    end
+    if $ui.yesno("\nIs the page reading order scanned left-to-right?", 'y')
+      ymltext << "reading_order: left-to-right\n"
+    else
+      ymltext << "reading_order: right-to-left\n"
+    end
+
     languages = Array.new
     if $ui.yesno("\nDoes the text include languages other than English?", 'n')
       $ui.message("\nEnter the 3-letter ISO code for each language in the text, including 'eng' if appropriate")
       $ui.message("Enter one language code at a time; enter blank when finished")
-      l = $ui.question('', 'Next language: ')
+      l = $ui.question('', 'First language: ')
       while l != ''
         languages << l
         l = $ui.question('', 'Next language: ')
@@ -150,6 +214,37 @@ class BatchHathiTrust < Batch
     else
       languages << "eng"
     end
+    if $ui.yesno("\nDoes the text include any script besides Latin script?", 'n')
+      $ui.message("\nEnter the name of each script starting with a capital letter, e.g. 'Fraktur', 'Cyrillic', 'Greek'")
+      $ui.message("Include 'Latin' if Latin script is also present")
+      $ui.message("Enter one language code at a time; enter blank when finished")
+      script = $ui.question('', 'First script: ')
+      while script != ''
+        languages << 'script/' + script
+        script = $ui.question('', 'Next script: ')
+      end
+    end
+
+    include_forms = $ui.yesno("\nWould you like to fill out a DASI and/or Coversheet for this submission?", 'y')
+
+    $ui.message('Creating package directory')
+    @copy_of_directory = @user_directory + '_Packaged'
+    if Dir.exist?(@copy_of_directory) then FileUtils.remove_dir(@copy_of_directory) end
+    FileUtils.mkdir_p @copy_of_directory
+    $ui.debug()
+
+    $ui.message("\nCopying images:")
+    n = 0
+    imagefiles.sort.each { |f|
+      n = n+1
+      newfile = File.join(@copy_of_directory, sprintf("%08d.tif", n))
+      FileUtils.cp(f, newfile)
+      $ui.message File.basename(f) + " --> " + File.basename(newfile)
+    }
+    Dir.chdir(@copy_of_directory)
+
+    # Create OCR files for each page
+    $ui.message("\nRunning OCR on images:")
     Dir.glob('*.tif').each { |f|
       filebase = File.basename(f, '.tif')
       command = "tesseract " + f + " " + filebase + " -l " + languages.join("+") + " hocr"
@@ -162,38 +257,13 @@ class BatchHathiTrust < Batch
     }
 
     # Create YAML meta file
-    $ui.message('Creating YAML file')
-    exif = EXIFR::TIFF.new('00000001.tif')
+    $ui.message("\nCreating YAML file")
     yml = File.open("meta.yml", "w+")
-    yml.puts "scanner_user: Tufts University, Tisch Library Digital Initiatives Dept."
-    if exif.date_time and exif.date_time != ""
-	  $ui.message "EXIF date_time: " + exif.date_time.to_s
-      yml.puts "capture_date: " + exif.date_time.strftime("%Y-%m-%dT%H:%M:%S%:z")
-    elsif
-      digidate = $ui.question("What day was this item digitized (mm/dd/yyyy)?")
-      datetime = Time.strptime(digidate, "%m/%d/%Y")
-	  yml.puts "capture_date: " + datetime.strftime("%Y-%m-%dT%H:%M:%S%:z")
-    end
-    if exif.make
-      yml.puts "scanner_make: " + exif.make
-    end
-    if exif.model
-      yml.puts "scanner_model: " + exif.model
-    end
-    if $ui.yesno("\nWas the resource scanned left-to-right?", 'y')
-      yml.puts "scanning_order: left-to-right"
-    else
-      yml.puts "scanning_order: right-to-left"
-    end
-    if $ui.yesno("\nIs the page reading order scanned left-to-right?", 'y')
-      yml.puts "reading_order: left-to-right"
-    else
-      yml.puts "reading_order: right-to-left"
-    end
+	yml.puts ymltext
     yml.close
 
     # Create MD5 checksum file
-    $ui.message("\nGenerating MD5 checksums")
+    $ui.message('Generating MD5 checksums')
     File.open("checksum.md5", "w+") { |c|
       Dir.glob(['*.tif', '*.txt', '*.html', 'meta.yml']).each { |f|
         hex = Digest::MD5.file(f).hexdigest
@@ -203,6 +273,7 @@ class BatchHathiTrust < Batch
 
     zipname = id.gsub(":", "+").gsub("/", "=") + '.zip'
     $ui.message('Creating zipfile')
+    Zip.write_zip64_support = true
     Zip::File.open(zipname, create: true) { |z|
       Dir.glob(['*.tif', '*.txt', '*.html', 'meta.yml', 'checksum.md5']).each { |f|
         z.add(f, f)
@@ -213,17 +284,62 @@ class BatchHathiTrust < Batch
       File.delete(f)
     }
 
-    include_forms = $ui.yesno("\nWould you like to fill out a DASI and/or Coversheet for this submission?", 'y')
-    if include_forms
-      f = File.join($xslt_path, 'HathiTrust_DASI.docx')
-      newfile = File.join(@copy_of_directory, "DASI_" + id + ".docx")
-      FileUtils.cp(f, newfile)
-      f = File.join($xslt_path, 'HathiTrust_Coversheet.docx')
-      newfile = File.join(@copy_of_directory, "Coversheet_" + id + ".docx")
-      FileUtils.cp(f, newfile)
+    if batch
+      $ui.message("\nMoving submission to batch submission directory")
+      if !File.exist?(batch_directory)
+        Dir.mkdir(batch_directory)
+      end
+      if include_forms
+        f = File.join($xslt_path, 'HathiTrust_DASI.docx')
+        newfile = File.join(batch_directory, "DASI_" + File.basename(batch_directory) + ".docx")
+        if !File.exist?(newfile)
+          $ui.message('Copying blank DASI')
+          FileUtils.cp(f, newfile)
+        end
+        f = File.join($xslt_path, 'HathiTrust_Coversheet.docx')
+        newfile = File.join(batch_directory, "Coversheet_" + File.basename(batch_directory) + ".docx")
+        if !File.exist?(newfile)
+          $ui.message('Copying blank Coversheet')
+          FileUtils.cp(f, newfile)
+        end
+      end
+      File.rename(zipname, File.join(batch_directory, zipname))
+      Dir.chdir(batch_directory)
+      if File.exist?(batch_xml)
+        $ui.message("Merging new MARC record with existing records in batch.")
+        tempfile = File.join(batch_directory, "temp.xml")
+        File.rename(batch_xml, tempfile)
+        f = File.open(tempfile)
+        batchwriter = MARC::XMLWriter.new(batch_xml)
+        reader = MARC::XMLReader.new(f, parser: "nokogiri")
+        reader.each do |batchrecord|
+   	      batchwriter.write(batchrecord)
+        end
+        f.close()
+        File.delete(tempfile)
+      else
+        batchwriter = MARC::XMLWriter.new(batch_xml)
+      end
+      batchwriter.write(record)
+      batchwriter.close
+      Dir.rmdir(@copy_of_directory)
+      $ui.message("\nThe submission package is in\n  " + batch_directory)
+    else
+      $ui.message('Copying MARC file in MARCXML format')
+      xmlname = File.join(@copy_of_directory, id.gsub(":", "+").gsub("/", "=") + '.xml')
+      MARC::XMLWriter.new(xmlname) { |w| w.write(record) }
+      if include_forms
+        $ui.message('Copying blank DASI and Coversheet')
+        f = File.join($xslt_path, 'HathiTrust_DASI.docx')
+        newfile = File.join(@copy_of_directory, "DASI_" + id + ".docx")
+        FileUtils.cp(f, newfile)
+        f = File.join($xslt_path, 'HathiTrust_Coversheet.docx')
+        newfile = File.join(@copy_of_directory, "Coversheet_" + id + ".docx")
+        FileUtils.cp(f, newfile)
+      end
+      $ui.message("\nThe submission package is in\n  " + @copy_of_directory)
     end
 
-    $ui.message("\nThe submission package is in\n  " + @copy_of_directory)
     if include_forms
       $ui.message("\nFill out and submit the DASI and Coversheet before submitting the MARC file and Zip file.\n")
     end
